fix: change `usePrevious()` custom hook to use Healthy Church version
Gonna see if this clears up my state bugs.
The version from https://github.com/sergeyleschev/react-custom-hooks/blob/main/src/hooks/usePrevious/usePrevious.js seems to be doing something different, and was not behaving as I expected it to.  AFAIK, the one in HC does what it's supposed to, so let's see if that works here.